+++ src/main/java/jp/yoshikipom/runlogapi/infra/record/RecordJpaRepository.java
@@ -1,4 +1,4 @@
-package jp.yoshikipom.runlogapi.infra;
+package jp.yoshikipom.runlogapi.infra.record;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
